+++ se_invoice/src/Plugin/ExtraField/Display/CustomerStatistics.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Drupal\se_invoice\Plugin\ExtraField\Display;
+
+use Drupal\Core\Entity\ContentEntityInterface;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\extra_field\Plugin\ExtraFieldDisplayFormattedBase;
+
+/**
+ * Example Extra field with formatted output.
+ *
+ * @ExtraFieldDisplay(
+ *   id = "invoice_statistics",
+ *   label = @Translation("Invoice statistics"),
+ *   bundles = {
+ *     "node.*",
+ *   }
+ * )
+ */
+class CustomerStatistics extends ExtraFieldDisplayFormattedBase {
+
+  use StringTranslationTrait;
+
+  public function getLabel() {
+    return $this->t('Invoice statistics');
+  }
+
+  public function getLabelDisplay() {
+    return 'above';
+  }
+
+  public function viewElements(ContentEntityInterface $entity) {
+    if (!$block = \Drupal::service('plugin.manager.block')
+      ->createInstance('invoice_statistics_customer', [])
+      ->build()) {
+      return [];
+    }
+
+    return [
+      ['#markup' => render($block)]
+    ];
+  }
+
+}